+++ CHANGELOG.md
@@ -1,7 +1,12 @@
-## 0.1.0 - 23/07/2019
+## 0.2.0 - 23/07/2019
 
-- Develop slider indicator
+- Develop new parameter for the slider (**Breaking**)
+- Now the parameter accept child instead of icon
 
 ## 0.1.1 - 23/07/2019
 
 - Add more description
+
+## 0.1.0 - 23/07/2019
+
+- Develop slider indicator